// src/app/calendar-core/models/enums.ts
export enum UpdateDetailsKind {
    Title = 'TITLE',
    Location = 'Location',
    Start = 'STRAT',
    End = 'END',
    Repeat = 'REPEAT',
    SaveToCalendar = 'SAVE_TO_CALENDAR',
    From = 'FROM',
    To = 'TO',
    Reminder = 'REMINDER',
    ShowAs = 'SHOW_AS',
    Body = 'BODY',
    IsAllDay = 'IS_ALL_DAY',
    IsPrivate = 'IS_PRIVATE',
    RequestResponses = 'REQUEST_RESPONSES',
    OnlineMeeting = 'OnlineMeeting'
}